Rename "-modal.tsx" file to "-page.tsx"
It's not really a modal
