+++ client/my-sites/domains/domain-management/index.jsx
@@ -7,7 +7,7 @@ import AddDnsRecord from './dns/add-dns-record';
 import DnsRecords from './dns/dns-records';
 import DomainConnectMapping from './domain-connect-mapping';
 import EditContactInfoPage from './edit-contact-info-page';
-import BulkEditContactInfoPage from './edit-contact-info-page/bulk-edit-contact-info-modal';
+import BulkEditContactInfoPage from './edit-contact-info-page/bulk-edit-contact-info-page';
 import ManageConsent from './manage-consent';
 import Security from './security';
 import Settings from './settings';
